requirements.dev.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
pydantic==0.32.2
PyMySQL==0.9.3
PyYAML==5.1.2
requests==2.22.0
urllib3==1.25.6
